Change stored min. thumbnail size as int not float

+++ PreviewText/Constants.swift
@@ -94,9 +94,9 @@ struct BUFFOON_CONSTANTS {
     }
 
     // FROM 1.0.5
-    static let MIN_THUMB_SIZE                   = 48.0
+    static let MIN_THUMB_SIZE                   = 48
     static let BASE_THUMB_LINE_SPACING          = 1.15
-    static let THUMB_SIZES: [CGFloat]           = [256.0, 128.0, 96.0, 64.0, 48.0, 32.0, 16.0]
+    static let THUMB_SIZES: [Int]               = [256, 128, 96, 64, 48, 32, 16]
 
     struct PREFS_IDS {
